feat: config setting to skip mapping of next statement only
When the proxy sees the following command from a client:

```sql
SET LOCAL UNSAFE_SKIP_MAPPING_NEXT_STATEMENT = t;
```

Parsing & mapping of the next statement ONLY will be skipped.

This is useful for working around `sqlparser` SQL grammar coverage
issues for database migrations in particular.

This command is not yet documented. It's possible that its full
implications may not be fully understood re: security and therefore it's
probably best to reveal its existence to customers on an as-needed
basis.

We should also consider a Proxy configuration option for completely
disabling this command.

NOT YET IMPLEMENTED:

- Setting this variable should also disable decryption of any results
  returned from the query, closing off a potential exploit.

- Integration test