ref(eql-mapper): remove extra Type implementation
Prior to this change the result of a successful type check was defined
in terms of a mirror of the `Unifier`'s `Type` enum but without type
variables or associated types: because all types would be resolved and
it theoretically simplified the consuming code.

Turns out it doesn't make any real difference so now we just export the
same `Type` representation that the `Unifier` uses.